day7_aws-cloud-automation-with-python-boto3/requirements.txt
awscrt==0.23.8
boto==2.49.0
boto3==1.37.15
botocore==1.37.15
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
jmespath==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.11.4
six==1.17.0
urllib3==2.3.0
